# internal/infra/webserver/handlers/response.go
package handlers

import (
	"encoding/json"
	"net/http"
)

type Response struct{}

func (*Response) JSON(w http.ResponseWriter, statusCode int, data interface{}) {
	if data == nil {
		w.WriteHeader(statusCode)
		return
	}

	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(statusCode)
	json.NewEncoder(w).Encode(data)
}